pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ng_loba"
version = "0.1.0"
description = "Line of Business Application components for NiceGUI"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "nicegui>=1.4.0",
    "tortoise-orm>=0.20.0",
    "pytz",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
addopts = "-W ignore::DeprecationWarning -W ignore::PendingDeprecationWarning -W ignore::pytest.PytestDeprecationWarning"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
